// src/types/schema.ts
export type Selector = string | string[];
export type FrameSelector = number[];

export enum SelectorType {
  CSS = "css",
  ARIA = "aria",
  Text = "text",
  XPath = "xpath",
}

export enum StepType {
  Change = "change",
  Click = "click",
  Close = "close",
  CustomStep = "customStep",
  DoubleClick = "doubleClick",
  EmulateNetworkConditions = "emulateNetworkConditions",
  Hover = "hover",
  KeyDown = "keyDown",
  KeyUp = "keyUp",
  Navigate = "navigate",
  Scroll = "scroll",
  SetViewport = "setViewport",
  WaitForElement = "waitForElement",
  WaitForExpression = "waitForExpression",
}

export enum AssertedEventType {
  Navigation = "navigation",
}

export interface NavigationEvent {
  type: AssertedEventType.Navigation;
  url?: Pattern;
  title?: Pattern;
}

export type AssertedEvent = NavigationEvent;

export interface BaseStep {
  type: StepType;
  timeout?: number;
  assertedEvents?: AssertedEvent[];
}

export interface StepWithTarget extends BaseStep {
  /**
   * Defaults to main
   */
  target?: Target;
}

export interface StepWithFrame extends StepWithTarget {
  /**
   * Defaults to main frame
   */
  frame?: FrameSelector;
}

export interface StepWithSelectors extends StepWithFrame {
  /**
   * A list of alternative selectors that lead to selection of a single element
   * to perform the step on. Currently, we support CSS selectors, ARIA selectors
   * (start with 'aria/'), XPath selectors (start with `xpath/`) and text
   * selectors (start with `text/`). Each selector could be a string or an array
   * of strings. If it's a string, it means that the selector points directly to
   * the target element. If it's an array, the last element is the selector for
   * the target element and the preceding selectors point to the ancestor
   * elements. If the parent element is a shadow root host, the subsequent
   * selector is evaluated only against the shadow DOM of the host (i.e.,
   * `parent.shadowRoot.querySelector`). If the parent element is not a shadow
   * root host, the subsequent selector is evaluated in the regular DOM (i.e.,
   * `parent.querySelector`).
   *
   * During the execution, it's recommended that the implementation tries out
   * all of the alternative selectors to improve reliability of the replay as
   * some selectors might get outdated over time.
   */
  selectors: Selector[];
}
// ...
export interface DoubleClickStep extends ClickAttributes, StepWithSelectors {
  type: StepType.DoubleClick;
}

export interface ClickStep extends ClickAttributes, StepWithSelectors {
  type: StepType.Click;
  /**
   * Delay (in ms) between the mouse up and mouse down of the click. Defaults to
   * 50ms.
   */
  duration?: number;
}

export interface HoverStep extends StepWithSelectors {
  type: StepType.Hover;
}

export interface ChangeStep extends StepWithSelectors {
  type: StepType.Change;
  value: string;
}

export interface EmulateNetworkConditionsStep extends StepWithTarget {
  type: StepType.EmulateNetworkConditions;
  download: number;
  upload: number;
  latency: number;
}

export interface KeyDownStep extends StepWithTarget {
  type: StepType.KeyDown;
  key: Key;
}

export interface KeyUpStep extends StepWithTarget {
  type: StepType.KeyUp;
  key: Key;
}

export interface CloseStep extends StepWithTarget {
  type: StepType.Close;
}

export interface SetViewportStep extends StepWithTarget {
  type: StepType.SetViewport;
  width: number;
  height: number;
  deviceScaleFactor: number;
  isMobile: boolean;
  hasTouch: boolean;
  isLandscape: boolean;
}

export interface ScrollPageStep extends StepWithFrame {
  type: StepType.Scroll;
  /**
   * Absolute scroll x position in px. Defaults to 0
   */
  x?: number;
  /**
   * Absolute scroll y position in px. Defaults to 0
   */
  y?: number;
}

export type ScrollElementStep = ScrollPageStep & StepWithSelectors;

export type ScrollStep = ScrollPageStep | ScrollElementStep;

export interface NavigateStep extends StepWithTarget {
  type: StepType.Navigate;
  url: string;
}

export interface CustomStepParams {
  type: StepType.CustomStep;
  name: string;
  parameters: unknown;
}

export type CustomStep =
  | (CustomStepParams & StepWithTarget)
  | (CustomStepParams & StepWithFrame);

export type UserStep =
  | ChangeStep
  | ClickStep
  | HoverStep
  | CloseStep
  | CustomStep
  | DoubleClickStep
  | EmulateNetworkConditionsStep
  | KeyDownStep
  | KeyUpStep
  | NavigateStep
  | ScrollStep
  | SetViewportStep;

/**
 * `waitForElement` allows waiting for the presence (or absence) of the number
 * of elements identified by the selector.
 *
 * For example, the following step would wait for less than three elements
 * to be on the page that match the selector `.my-class`.
 *
 * ```
 * {
 *   "type": "waitForElement",
 *   "selectors": [".my-class"],
 *   "operator": "<=",
 *   "count": 2,
 * }
 * ```
 */
export interface WaitForElementStep extends StepWithSelectors {
  type: StepType.WaitForElement;
  /**
   * @defaultValue `'=='`
   */
  operator?: ">=" | "==" | "<=";
  /**
   * @defaultValue `1`
   */
  count?: number;
  /**
   * Whether to wait for elements matching this step to be hidden. This can be
   * thought of as an inversion of this step.
   *
   * @defaultValue `true`
   */
  visible?: boolean;
  /**
   * Whether to also check the element(s) for the given properties.
   */
  properties?: Partial<JSONSerializable<HTMLElement>> & {
    [key: string]: JSONValue;
  };
  /**
   * Whether to also check the element(s) for the given attributes.
   */
  attributes?: { [name: string]: string };
}

/**
 * `waitForExpression` allows for a JavaScript expression to resolve to truthy
 * value.
 *
 * For example, the following step pauses for two seconds and then resolves to
 * true allowing the replay to continue.
 *
 * ```
 * {
 *   "type": "waitForExpression",
 *   "expression": "new Promise(resolve => setTimeout(() => resolve(true),
 * 2000))",
 * }
 * ```
 */
export interface WaitForExpressionStep extends StepWithFrame {
  type: StepType.WaitForExpression;
  expression: string;
}

export type AssertionStep = WaitForElementStep | WaitForExpressionStep;

export type Step = UserStep | AssertionStep;